rename
wip mem-133 `Integration test(android)` が成功するようにする
